% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_utils.R
\name{pop_parse_filenames}
\alias{pop_parse_filenames}
\title{Parse metadata from multiple WorldPop raster filenames}
\usage{
pop_parse_filenames(file_paths)
}
\arguments{
\item{file_paths}{Character vector. Paths to WorldPop raster files.}
}
\value{
Tibble with one row per file and the same columns as
\code{\link[=pop_parse_filename]{pop_parse_filename()}}.
}
\description{
Vectorised wrapper around \code{\link[=pop_parse_filename]{pop_parse_filename()}}.
}
